Move lifecycle kotlin bits to AutoDisposeLifecycle.kt
